requirements.txt
Flask
Flask-Cors
google-generativeai
pdfplumber
spacy
python-dotenv
langchain-huggingface
gunicorn
python-docx